requirements.txt
winrt-runtime
winrt-Windows.Devices.Sensors
Pillow
